design(process): split the Organized Wireframe out from Mockup
In order to make things much more clear, I have decided to split
the Organized Wireframe up from the Mockup as its own deliverable.
Previously, it was simply a step in the process of creating a
Mockup. However, I think that having it exist as its own deliverable
is nicer because it creates a clear separation between the two.

Furthermore, it makes demonstrating the process to others much
simpler instead of asking them to imagine something they can just
use the name "Organized Wireframe" to refer to this part of the
process. In other words, there is power in names!

I have also created some more icons during my morning design
practice, as part of the TechCrunch tracing project.